requirements.txt
streamlit>=1.39.0
openai>=1.51.0
python-dotenv>=1.0.1
pandas>=2.2.0
